requirements.txt
numpy>=1.24,<3.0
pandas>=2.0,<3.0
matplotlib>=3.7,<4.0
seaborn>=0.13,<1.0
scikit-learn>=1.3,<2.0
mlflow>=2.12,<3.0
jupyter>=1.0,<2.0
ipykernel>=6.0,<7.0
optuna
mlflow
hydra-core
scikit-learn
pandas
numpy
joblib
pytest
PyYAML
black==23.12.1
flake8==7.3.0